Fixed Atom and AtomProperty
Atom is now more than just _ulong.
It is a ctype that can take AtomProperty or _ulong.
AtomProperty symbols are converted to _ulong using the _enum.
However they are not converted back to symbols when coming from C, and stay as a _ulong.
I'm not sure that's really problematic though.